Rearrange folder in settings
