--- maisim/maisim.Game/Graphics/UserInterface/Overlays/DebugSection.cs
@@ -1,19 +1,0 @@
-using osu.Framework.Graphics;
-using osu.Framework.Localisation;
-
-namespace maisim.Game.Graphics.UserInterface.Overlays
-{
-    public class DebugSection : SettingsSection
-    {
-        public override bool EnableSeperator => true;
-        public override LocalisableString Header => "debug";
-
-        public DebugSection()
-        {
-            Children = new Drawable[]
-            {
-                new FolderSettings()
-            };
-        }
-    }
-}
