+++ maisim/maisim.Game/Graphics/UserInterface/Overlays/Settings/AudioSection.cs
@@ -1,3 +1,4 @@
+using maisim.Game.Graphics.UserInterface.Overlays.Settings.Audio;
 using osu.Framework.Graphics;
 using osu.Framework.Localisation;
